GH-25: Setup base test class for extension
